Fix parse_color differences
