+++ pyvista/plotting/renderer.py
@@ -12,7 +12,8 @@ from vtk import vtkRenderer
 import pyvista
 from pyvista.utilities import wrap
 
-from .theme import parse_color, parse_font_family, rcParams, MAX_N_COLOR_BARS
+from .colors import parse_color
+from .theme import parse_font_family, rcParams, MAX_N_COLOR_BARS
 from .tools import create_axes_marker
