+++ pyvista/plotting/theme.py
@@ -110,33 +110,6 @@ def set_plot_theme(theme):
             rcParams[k] = v
 
 
-
-def parse_color(color, opacity=None):
-    """Parses color into a vtk friendly rgb list.
-    Values returned will be between 0 and 1.
-    """
-    if color is None:
-        color = rcParams['color']
-    if isinstance(color, str):
-        color = string_to_rgb(color)
-    elif len(color) == 3:
-        pass
-    elif len(color) == 4:
-        color = color[:3]
-    else:
-        raise Exception("""
-    Invalid color input: ({})
-    Must ba string, rgb list, or hex color string.  For example:
-        color='white'
-        color='w'
-        color=[1, 1, 1]
-        color='#FFFFFF'""".format(color))
-    if opacity is not None and isinstance(opacity, (float, int)):
-        color = [color[0], color[1], color[2], opacity]
-    return color
-
-
-
 def parse_font_family(font_family):
     """ checks font name """
     # check font name
